move AddrToString to hd_wallet/address_ops.rs
