+++ mm2src/coins/hd_wallet/address_ops.rs
@@ -6,6 +6,16 @@ pub trait DisplayAddress {
     fn display_address(&self) -> String;
 }
 
+/// Should convert coin `Self::Address` type into a properly formatted string representation.
+///
+/// Don't use `to_string` directly on `Self::Address` types in generic TPU code!
+/// It may produce abbreviated or non-standard formats (e.g. `ethereum_types::Address` will be like this `0x7cc9…3874`),
+/// which are not guaranteed to be parsable back into the original `Address` type.
+/// This function should ensure the resulting string is consistently formatted and fully reversible.
+pub trait AddrToString {
+    fn addr_to_string(&self) -> String;
+}
+
 /// `HDAddressOps` Trait
 ///
 /// Defines operations associated with an HD (Hierarchical Deterministic) address.
